no more duplicate

+++ frontend/src/views/PostView.ts
@@ -67,25 +67,21 @@ export class PostView extends Backbone.View<any> {
     private deleteLikePV(e) {
         const postId = $(e.currentTarget).attr("data-id");
         const that = this;
-        const likeToDestroy = _.findWhere(this.model.likes, {user : HeaderRequestGenerator.currentUser()}) as LikeModel;
-        const myNewLike = new LikeModel(likeToDestroy);
-        console.log(likeToDestroy);
+        if (this.model.id.toString() === postId) {
+            const likeToDestroy = _.findWhere(this.model.likes, {user : HeaderRequestGenerator.currentUser()}) as LikeModel;
+            const myNewLike = new LikeModel(likeToDestroy);        
+            myNewLike.destroy({
+                beforeSend: HeaderRequestGenerator.sendAuthorization,
+                success() {
+                    that.updateLikesCountTemp(false, postId, e);
+                },
+                error() {
+                    // TODO handle error
+                }
+            });
+        }
         
-        /*if(likeToDestroy != undefined) {
-            var theId = likeToDestroy.id;
-            var myLike = new LikeModel({id:theId,pictureId:likeToDestroy.pictureId,userId:likeToDestroy.user});
-            console.log(myLike);
-            console.log(likeToDestroy);
-        }*/
-        myNewLike.destroy({
-            beforeSend: HeaderRequestGenerator.sendAuthorization,
-            success() {
-                that.updateLikesCountTemp(false, postId, e);
-            },
-            error() {
-                // TODO handle error
-            }
-        });     
+        
     }
 
     private addComment(e) {
